Issue #11446: Updated testTabViolation in Checker Test
