+++ src/test/java/com/puppycrawl/tools/checkstyle/bdd/TestInputConfiguration.java
@@ -58,6 +58,8 @@ public final class TestInputConfiguration {
             "com.puppycrawl.tools.checkstyle.checks.whitespace.FileTabCharacterCheck",
             "com.puppycrawl.tools.checkstyle.api.AbstractFileSetCheckTest$ViolationFileSetCheck",
             "com.puppycrawl.tools.checkstyle.api.FileSetCheckTest$TestFileSetCheck",
+            "com.puppycrawl.tools.checkstyle.internal.testmodules"
+                + ".VerifyPositionAfterLastTabFileSet",
             "com.puppycrawl.tools.checkstyle.CheckerTest$VerifyPositionAfterTabFileSet"
     ));
