+++ src/test/java/com/puppycrawl/tools/checkstyle/internal/testmodules/VerifyPositionAfterLastTabFileSet.java
@@ -1,0 +1,42 @@
+///////////////////////////////////////////////////////////////////////////////////////////////
+// checkstyle: Checks Java source code and other text files for adherence to a set of rules.
+// Copyright (C) 2001-2024 the original author or authors.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+///////////////////////////////////////////////////////////////////////////////////////////////
+
+package com.puppycrawl.tools.checkstyle.internal.testmodules;
+
+import java.io.File;
+
+import com.puppycrawl.tools.checkstyle.api.AbstractFileSetCheck;
+import com.puppycrawl.tools.checkstyle.api.FileText;
+
+public class VerifyPositionAfterLastTabFileSet extends AbstractFileSetCheck {
+
+    @Override
+    protected void processFiltered(File file, FileText fileText) {
+        int lineNumber = 0;
+        for (String line : getFileContents().getLines()) {
+            final int position = line.lastIndexOf('\t');
+            lineNumber++;
+
+            if (position != -1) {
+                log(lineNumber, position + 1, "violation");
+            }
+        }
+    }
+
+}
