--- src/test/resources/com/puppycrawl/tools/checkstyle/checker/InputCheckerTabCharacter.txt
@@ -1,13 +1,0 @@
-/*
-com.puppycrawl.tools.checkstyle.CheckerTest$VerifyPositionAfterTabFileSet
-
-*/
-
-// has no tabs
-line1
-
-// has a tab
-    line4	// violation
-
-// has 2 tabs
-        line5	// violation	// violation
